@@ -16,13 +16,14 @@
         }
     ],
     "require": {
-        "php": "^7.2|^8.0",
-        "illuminate/support": "^6.0|^7.0|^8.0|^9.0|^10.0",
-        "livewire/livewire": "^2.0"
+        "php": "^8.1",
+        "illuminate/support": "^10.0",
+        "livewire/livewire": "^v3.0.0-beta.7"
     },
     "require-dev": {
-        "orchestra/testbench": "^5.0|^6.0",
-        "phpunit/phpunit": "^8.0|^9.0"
+        "laravel/pint": "^1.10",
+        "orchestra/testbench": "^8.0",
+        "phpunit/phpunit": "^10.1"
     },
     "autoload": {
         "psr-4": {
@@ -46,10 +47,7 @@
         "laravel": {
             "providers": [
                 "Asantibanez\\LivewireCalendar\\LivewireCalendarServiceProvider"
-            ],
-            "aliases": {
-                "LivewireCalendar": "Asantibanez\\LivewireCalendar\\LivewireCalendarFacade"
-            }
+            ]
         }
     }
 }
